requirements.txt
flask 
google-cloud-aiplatform===1.73.0
google-cloud-spanner 
